Separation of Concerns & AJAX Calls
I have separated my html and css folder, as the project was becoming
somewhat unwieldy. I also hooked up JavaScript and jQuery to the html
file, and wrote an ajax call function for every query the user submits.
In addition, I created the elements in the html that will be displaying
the weather conditions.